Fix URL path issue in Windows. See issue #68
We were using path.join for joining url paths, which obviously
failed in Windows since url paths don't work like win fs paths.
There is no equiv url.join (url.resolve is a little different)
so for now this is the simplest fix to get it to work. The paths
used are within our control via the resource modules so I
think this is ok.