Check if a callback is passed to Retrieve Customer
If a parameter is passed to Retrieve customer that isn't an
auth key, the library incorrectly assumed it was a account id.

Updated the check to ensure that a function could be passed in as
a sole argument and it would retrieve the account for the pre-set
api key.

Fixed #147