Add instructions for running single tests
Adds a section in the README like some of our other languages for
running either a single file of tests or a single test within a file. A
script is also added to `package.json` to support this.

Slightly unfortunate: the `--grep` flag of Mocha is case sensitive which
will probably lead to confusion because we use a lot of inconsistent
casing in test names. There is a case insensitive version in modern
version of Mocha, but we'll have to upgrade to 3.0.0+ to get it.