Improve error message when ommitting a parameter
Improve the error message when a `stripeMethod` is called but a
parameter is accidentally ommitted. Previously, we were accidentally
including the product of `utils.makeURLInterpolator` in the error
string, which made it very difficult to read:

```
Error: Stripe: Argument "id" required, but got: undefined (on API request to GET function (outputs) {
        return cleanString.replace(/\{([\s\S]+?)\}/g, function($0, $1) {
          return encodeURIComponent(outputs[$1] || '');
        });
      })
    at Object.<anonymous> (/Users/brandur/stripe/stripe-node/lib/StripeMethod.js:53:21)
```

I think this was basically an implementation accident, as what it's
actually trying to do is print the path of the API call that's about to
be invoked. We now produce this instead:

```
Error: Stripe: Argument "id" required, but got: undefined (on API request to `GET /invoices/{id}`)
    at Object.<anonymous> (/Users/brandur/stripe/stripe-node/lib/StripeMethod.js:53:21)
```

Fixes #323.